Add into docs.

+++ examples/executor/executor_lag.py
@@ -8,53 +8,66 @@ from pymoa.executor.threading import ThreadExecutor, AsyncThreadExecutor
 from pymoa.executor.remote.dummy import DummyRemoteExecutor
 from pymoa.device.analog import RandomAnalogChannel
 from pymoa.executor.remote.socket.websocket_client import WebSocketExecutor
+from pymoa.executor.remote.socket.multiprocessing_client import \
+    MultiprocessSocketExecutor
 
 
-async def measure_within_process_quart_lag(cls):
+async def measure_executor(executor, name):
+    cls = executor.__class__
+    device = RandomAnalogChannel(executor=executor)
+    responses = []
+
+    for _ in range(100):
+        ts, t, te = await executor.get_echo_clock()
+        responses.append((te - ts) / 1e6)
+
+    if hasattr(executor, 'ensure_remote_instance'):
+        await executor.ensure_remote_instance(device)
+
+    if not isinstance(executor, AsyncThreadExecutor):
+        ts = time.perf_counter_ns()
+        for _ in range(100):
+            await device.read_state()
+
+        te = time.perf_counter_ns()
+        rate = 100 * 1e9 / (te - ts)
+
+        await device.pump_state(100)
+        rate_cont = 100 * 1e9 / (time.perf_counter_ns() - te)
+    else:
+        rate = rate_cont = 0
+
+    if hasattr(executor, 'delete_remote_instance'):
+        await executor.delete_remote_instance(device)
+
+    response = sum(responses) / len(responses)
+    print(f'{name} - {cls.__name__}; '
+          f'Round-trip lag: {response:.2f}ms. '
+          f'Rate: {rate:.2f}Hz. '
+          f'Continuous rate: {rate_cont:.2f}Hz')
+
+
+async def measure_within_process_quart_lag(cls, port):
     app = create_app()
     async with trio.open_nursery() as socket_nursery:
         if cls is RestExecutor:
-            executor = RestExecutor(uri='http://127.0.0.1:5001')
+            executor = RestExecutor(uri=f'http://127.0.0.1:{port}')
         else:
             executor = WebSocketExecutor(
-                nursery=socket_nursery, server='127.0.0.1', port=5001)
-
-        device = RandomAnalogChannel(executor=executor)
-        requests = []
-        responses = []
+                nursery=socket_nursery, server='127.0.0.1', port=port)
 
         async with trio.open_nursery() as nursery:
-            nursery.start_soon(app.run_task, "127.0.0.1", 5001)
+            nursery.start_soon(app.run_task, "127.0.0.1", port)
             await trio.sleep(1)
 
             async with app.app_context():
                 await executor.start_executor()
 
-                for _ in range(100):
-                    ts, t, te = await executor.get_echo_clock()
-                    requests.append((t - ts) / 1e6)
-                    responses.append((te - t) / 1e6)
+                await measure_executor(executor, 'Quart-internal')
 
-                await executor.ensure_remote_instance(device)
-                ts = time.perf_counter_ns()
-                for _ in range(100):
-                    await device.read_state()
-                te = time.perf_counter_ns()
-
-                await device.pump_state(100)
-                rate_cont = 100 * 1e9 / (time.perf_counter_ns() - te)
-
-                await executor.delete_remote_instance(device)
                 await executor.stop_executor(block=True)
 
             nursery.cancel_scope.cancel()
-
-    request = sum(requests) / len(requests)
-    response = sum(responses) / len(responses)
-    print(f'Quart-internal - {cls.__name__}; Request lag: {request:.2f}ms. '
-          f'Response lag: {response:.2f}ms. '
-          f'Rate: {100 * 1e9 / (te - ts):.2f}Hz. '
-          f'Continuous rate: {rate_cont:.2f}Hz')
 
 
 async def measure_outside_process_quart_lag(cls, host='127.0.0.1', port=5000):
@@ -65,74 +78,34 @@ async def measure_outside_process_quart_lag(cls, host='127.0.0.1', port=5000):
             executor = WebSocketExecutor(
                 nursery=socket_nursery, server=host, port=port)
 
-        device = RandomAnalogChannel(executor=executor)
-        responses = []
-
         await executor.start_executor()
         await trio.sleep(1)
 
-        for _ in range(100):
-            ts, t, te = await executor.get_echo_clock()
-            responses.append((te - ts) / 1e6)
-
-        await executor.ensure_remote_instance(device)
-        ts = time.perf_counter_ns()
-        for _ in range(100):
-            await device.read_state()
-        te = time.perf_counter_ns()
-
-        await device.pump_state(100)
-        rate_cont = 100 * 1e9 / (time.perf_counter_ns() - te)
-        await executor.delete_remote_instance(device)
+        await measure_executor(executor, 'Quart-external')
 
         await executor.stop_executor(block=True)
 
-    response = sum(responses) / len(responses)
-    print(f'Quart-external - {cls.__name__}; '
-          f'Round-trip lag: {response:.2f}ms. '
-          f'Rate: {100 * 1e9 / (te - ts):.2f}Hz. '
-          f'Continuous rate: {rate_cont:.2f}Hz')
 
-
-async def measure_cls_lag(cls):
-    executor = cls()
-    device = RandomAnalogChannel(executor=executor)
-    requests = []
-    responses = []
+async def measure_multiprocess_lag(host='127.0.0.1', port=5000):
+    executor = MultiprocessSocketExecutor(server=host, port=port)
 
     await executor.start_executor()
     await trio.sleep(1)
 
-    for _ in range(100):
-        ts, t, te = await executor.get_echo_clock()
-        requests.append((t - ts) / 1e6)
-        responses.append((te - t) / 1e6)
-
-    if not isinstance(executor, AsyncThreadExecutor):
-        if hasattr(executor, 'ensure_remote_instance'):
-            await executor.ensure_remote_instance(device)
-
-        ts = time.perf_counter_ns()
-        for _ in range(100):
-            await device.read_state()
-        rate = 100 * 1e9 / (time.perf_counter_ns() - ts)
-
-        ts = time.perf_counter_ns()
-        await device.pump_state(100)
-        rate_cont = 100 * 1e9 / (time.perf_counter_ns() - ts)
-
-        if hasattr(executor, 'ensure_remote_instance'):
-            await executor.delete_remote_instance(device)
-    else:
-        rate_cont = rate = 0
+    await measure_executor(executor, 'multiprocess')
 
     await executor.stop_executor(block=True)
 
-    request = sum(requests) / len(requests)
-    response = sum(responses) / len(responses)
-    print(f'{cls.__name__}; Request lag: {request:.2f}ms. '
-          f'Response lag: {response:.2f}ms. Rate: {rate:.2f}Hz. '
-          f'Continuous rate: {rate_cont:.2f}Hz')
+
+async def measure_cls_lag(cls):
+    executor = cls()
+
+    await executor.start_executor()
+    await trio.sleep(1)
+
+    await measure_executor(executor, 'local')
+
+    await executor.stop_executor(block=True)
 
 
 async def measure_no_executor_lag():
@@ -158,12 +131,17 @@ if __name__ == '__main__':
         '--host', dest='host', action='store', default="127.0.0.1")
     parser.add_argument(
         '--port', dest='port', action='store', default=5000, type=int)
+    parser.add_argument(
+        '--port_internal', dest='port_internal', action='store', default=5001,
+        type=int)
     args = parser.parse_args()
 
     for cls in (RestExecutor, WebSocketExecutor):
-        trio.run(measure_within_process_quart_lag, cls)
+        trio.run(measure_within_process_quart_lag, cls, args.port_internal)
         # only run if quart app is serving externally
         trio.run(measure_outside_process_quart_lag, cls, args.host, args.port)
+
+    trio.run(measure_multiprocess_lag, args.host, args.port_internal)
 
     for cls in (ThreadExecutor, AsyncThreadExecutor, DummyRemoteExecutor):
         trio.run(measure_cls_lag, cls)
@@ -179,6 +157,8 @@ Quart-internal - WebSocketExecutor; Request lag: 0.63ms. \
 Response lag: 0.75ms. Rate: 324.92Hz. Continuous rate: 795.35Hz
 Quart-external - WebSocketExecutor; Round-trip lag: 2.20ms. \
 Rate: 533.46Hz. Continuous rate: 1104.19Hz
+multiprocess - MultiprocessSocketExecutor; Round-trip lag: 0.48ms. \
+Rate: 1132.06Hz. Continuous rate: 3144.60Hz
 ThreadExecutor; Request lag: 0.21ms. Response lag: 0.45ms. \
 Rate: 2789.56Hz. Continuous rate: 5395.46Hz
 AsyncThreadExecutor; Request lag: 0.65ms. Response lag: 0.51ms. Rate: 0.00Hz. \
